getErrorMessage moved to transferTracking
