+++ app/src/utils/snowbridge.ts
@@ -244,11 +244,3 @@ const getTransferTimestamp = (
     ? transferResult.info.when.getTime()
     : transferResult.originBlockTimestamp
 
-export function getErrorMessage(err: unknown) {
-  let message = 'Unknown error'
-  if (err instanceof Error) {
-    message = err.message
-  }
-  console.error(message, err)
-  return message
-}
