+++ app/src/utils/transferTracking.ts
@@ -1,0 +1,8 @@
+export function getErrorMessage(err: unknown) {
+    let message = 'Unknown error'
+    if (err instanceof Error) {
+        message = err.message
+    }
+    console.error(message, err)
+    return message
+}
